add ut for drain capture at coordinator level.
