+++ docs/swagger/swagger.json
@@ -73,6 +73,12 @@
                         "schema": {
                             "$ref": "#/definitions/model.HTTPError"
                         }
+                    },
+                    "503": {
+                        "description": "Service Unavailable",
+                        "schema": {
+                            "$ref": "#/definitions/model.HTTPError"
+                        }
                     }
                 }
             }
