+++ docs/swagger/swagger.yaml
@@ -282,6 +282,10 @@ paths:
           description: Internal Server Error
           schema:
             $ref: '#/definitions/model.HTTPError'
+        "503":
+          description: Service Unavailable
+          schema:
+            $ref: '#/definitions/model.HTTPError'
       summary: Drain captures
       tags:
       - capture
